fixed shop update issue
In shop.cpp, the Update function was skipping over items that are "socketed", which I believe any item that is found in this program should be available to be listed to be sold. Will have to dig further to see how it determines what is "Socketed", since it seems that everything suddenly became socketed.