Converted spaces to tabs
also added version.txt file to repo, to tell the program what the latest version is to other users using my fork's version of this program. The updater will check this file, and compare it to its current VER_CODE version, notifying them to update when version.txt is greater than VER_CODE in their installation.