items table cleanup and NULL checking
Sometimes (not sure how right now), there will be null keys when writing to the items table. The primary key in the items table is the stash/character id that contains items. For now I'm going to just have Acquisition cleanup the NULL records each time at launch, and also return when receiving a NULL location instead of writing to the database.

Can look into seeing how these NULL locations are popping up in the future, but figured to at least get this cleanup code in place first.