dynamic rate limiting
This is a rewrite of ItemsManagerWroker's internals, using a new RateLimiter class that implements dynamic rate limiting based on info from the HTTP response headers.

Removes the QNetworkDiskCache based tab caching scheme. Instead only checked or selected tabs are requested.

Adds a rate limit status panel to the main window. It's accessible from a Rate Limit Status button on the right of the status bar, and shows the internal state of the rate limit policy managers.

Bundles several other improvements:

Moved ParseItemMods into it's own threads because it was blocking the UI from updating the status bar during it's initial loading of those mods from the SQLITE data store.

Added a remove-only flag to ItemLocation. This was needed to stop propagating buyouts in remove-only tabs, because this was forcing those tabs to be refreshed. There was no way to turn those refreshes off, because it's not possible to change notes on items in remove-only tabs. GGG also changed the trade site so remove-only stash tabs are no longer public.

Also changed ItemLocation so it uses the unique tab id provided by GGG instead of the tab name when getting a unique tab hash or checking the validity of a tab. This is because stash tab labels are not guaranteed unique, and they can even be empty.

Bumped the version to 0.9.9-rc1.

(I skipped 0.9.8 because that was an aborted release that was bugged and causing a flood of API requests).