Deployment cleanup and changes
Added deploy-win.cmd that does everything after QT has made a release build, and update the Inno 6 installer script for building the installer on windows.

Removed deployment stuff that's not in use and would need to be updated:
- travis (Linux build using QT 5.9 and gcc 4.9)
- appveyor (used VS 2017 and QT 5.9)
- deploy.ps1 (used QT IFW 4.2)

I also removed deploy-win.sh. This looks like it was meant to be run on windows, but there's no detail.  Not sure if it used Cygwin or WSL, or what was needed to make this work.