Revert tab hashing to v0.9.7
Tab labels were used to hash tab locations, but tab labels can be changed by the user, set to an empty value, and are not guaranteed to be unique.

Instead, we should be using the unique ID from GGG, but the tab hash is used to store buyouts. That means if we just change the hash here buyouts created in previous versions of acquisition will be lost.

Need a way to migrate the tab buyouts safely. That won't be a simple fix.