Use one QNetworkAccessManager and UpdateChecker
Create a QNetworkAccessManager and UpdateChecker in the Application class.

Replace all other network access managers and update checkers with references to these new ones.

Remove a number of members from the ItemsManagerWorker that were just references to members of the Application object. Use that instead.

This paves the way for OAuth and full rate-limiting of all network requests, regardless of where they are made in the application.

The update checker was only moved for simplicity.