Refactor UI initialization
Use explicit calls to connect() instead of relying on Qt to connect actions implicitly based on method names.

Rename UI actions for better clarity.

Use the new global datastore to load and save the theme, so it will be available to the login dialog once a theme has been set.