Change database interface
Now storing tabs as individual rows with character name or unique id as the primary key.

Changed database table definitions. This is not backwards compatible, and there it not yet a way to import older data stores.

Added a helper function to datastore to deserialize a single tab.

Due to the previous improvements to the data store interface, these changes are transparent to the rest of the application.