Changed for the new API
Due to the new api endpoints, we can no longer determine the url for rate limiting solely from the network requests, because the url path may have parameters like <league> or <stash_id>.

Instead, we now force the calling application to pass a string that defines the endpoint into Submit()

Also, keep track of the policy name inside the PolicyManger now. This simplified the logic around getting the name from the pointer to the Policy, which can be null.