Add temporary default policies
As of 3.22.2, we can't use HEAD requests to check the rate limit status for endpoints against the new API. This has been confirmed to an oversight that will be changed in the next game patch.

RateLimiter is designed around automatically checking endpoints, but for now we are going to hard-code endpoints and policies.