Remove item & location hashing
Hashing makes buyouts difficult to track, and dependent on the precise item and stash location json.

Instead, use the unique identifiers provided by GGG in the Stash, Item, and Character objects.

This also we means we lose the ability migrate buyouts. This will be replaced with the ability to import buyouts from older data files.