Better selection handling
Now if you change the sort order, the selected item should be preserved.

If an item is selected when the search form is updated, and that item is present in the updated search results, it will remain selected. Otherwise, the current item tooltip info is removed.

This behavior is not preserved between search tabs.