Updated windows deployment with cmake to simplify versioning
Now meta data like version strings are all defined in one place, the top of the CMakeLists.txt input file. From there they are automatically propagated into the versions header file and windows installer script.

Previously the version information in the qmake input file had to be manually synched with the version headers, and this was checked with static asserts in main(). Those checks are no longer needed.

This build dir is still hard-coded into the deploy-win.cmd script, and must be matched to the compiler and Qt version being used. Currently this is Qt Creator with MSVC2019.

The windows installer script now does some basic sanity checking to make sure build and version information are defined.