Move serialization & deserialization into datastore
Previously item and tab serialization was left to the data store user and a data store would return raw json. Now handles Locations and Items objects.

These functions should be the responsibility of the data store to hide implementation details. Now we can pass in lists of items and locations (as const pointers), and let the datastore prepare these for storage or retrieve them as appropriate.