Refactor (3 of 3) - items manager worker
Making ItemsManagerWorker an abstract base class, so that we can have one subclass for the legacy API with POESESSID, and another subclass for the OAuth API.

NOTE: This is the third of three commits that probably won't work unless they are all included. I'm trying to break out the changes for understandability.