Refactored how the status bar is updated
Reduced the number of program states and added a string message argument to the update signal. This not puts the responsibility of the code emitting updates to prepare the status message.

I did this because I wanted to add more status updates, but that was going to mean adding more fields to the status structure, some of which would  only be used for some status messages.

Hopefully this is simpler in the long run.