Simplify rate limit status signaling
The rate limiter used to live with the items worker, so it took extra code to glue rate limit status updates to the main window.

Now the rate limiter lives in the application object, so we can connect it's signals directly to the main window and get rid of the glue signals and slots.