Regression: update search results when a modifier search is active
I introduced an optimization in v0.11.1-alpha.10 that keeps track of which filters are active, so that only active filters need to be used when filtering items. However, I forgot about mods filter, which wasn't setting it's active state, so it was always being treated as inactive. Now the mods filter considers itself active whenever there is at least one mod search.

Also fix a UI bug that was causing the "Add Mod" button to be displayed incorrectly at program startup.