Add StashTab type to ItemLocation
The "type" field is a member of the StashTab object, and can be used to determine if a tab is a special tab like a currency tab. In the case of a quad tab, the location on the minimap display needs to be calculated differently, since quad tabs have twice as many inventory slots as normal tabs.