Change timing to avoid rate limit violations
There are still some hard-to-reproduce rate limit violations, so I've made some timing changes to hopefully avoid them.  Now requests for a specific policy will not be sent more than 1 per second, and if a request needs to be paused to avoid a violation, it's paused for a few extra seconds more.